feat(step3): Dealer, Challenger 클래스 작성..
